+++ parse/ParseProcess.cpp
@@ -1,9 +1,9 @@
 #include "ParseProcess.hpp"
 
-std::vector<std::pair<Resource*, size_t>> ParseProcess::_required_resources;
-std::vector<std::pair<Resource*, size_t>> ParseProcess::_produced_resources;
+std::vector<std::pair<Resource*, double>> ParseProcess::_required_resources;
+std::vector<std::pair<Resource*, double>> ParseProcess::_produced_resources;
 
-void ParseProcess::AddRequiredResource(const std::string& name, size_t number, Graph* graph) {
+void ParseProcess::AddRequiredResource(const std::string& name, double number, Graph* graph) {
     if (graph->GetResourceByName(name) == nullptr) {
         graph->AddResource(Resource(name, 0));
     }
@@ -16,7 +16,7 @@ void ParseProcess::AddRequiredResource(const std::string& name, size_t number, G
     _required_resources.emplace_back(graph->GetResourceByName(name), number);
 }
 
-void ParseProcess::AddProducedResource(const std::string& name, size_t number, Graph* graph) {
+void ParseProcess::AddProducedResource(const std::string& name, double number, Graph* graph) {
     if (graph->GetResourceByName(name) == nullptr) {
         graph->AddResource(Resource(name, 0));
     }
@@ -34,10 +34,10 @@ void ParseProcess::CleanUp() {
     _produced_resources.clear();
 }
 
-std::vector<std::pair<Resource*, size_t>> ParseProcess::GetRequiredResources() {
+std::vector<std::pair<Resource*, double>> ParseProcess::GetRequiredResources() {
     return _required_resources;
 }
 
-std::vector<std::pair<Resource*, size_t>> ParseProcess::GetProducedResources() {
+std::vector<std::pair<Resource*, double>> ParseProcess::GetProducedResources() {
     return _produced_resources;
 }
