+++ parse/ParseProcess.hpp
@@ -4,15 +4,15 @@
 
 class ParseProcess {
  public:
-    static void AddRequiredResource(const std::string& name, size_t number, Graph* graph);
-    static void AddProducedResource(const std::string& name, size_t number, Graph* graph);
+    static void AddRequiredResource(const std::string& name, double number, Graph* graph);
+    static void AddProducedResource(const std::string& name, double number, Graph* graph);
 
-    static std::vector<std::pair<Resource*, size_t>> GetRequiredResources();
-    static std::vector<std::pair<Resource*, size_t>> GetProducedResources();
+    static std::vector<std::pair<Resource*, double>> GetRequiredResources();
+    static std::vector<std::pair<Resource*, double>> GetProducedResources();
 
     static void CleanUp();
 
  private:
-    static std::vector<std::pair<Resource*, size_t>> _required_resources;
-    static std::vector<std::pair<Resource*, size_t>> _produced_resources;
+    static std::vector<std::pair<Resource*, double>> _required_resources;
+    static std::vector<std::pair<Resource*, double>> _produced_resources;
 };
